update start to include sidekiq

+++ start.sh
@@ -5,9 +5,14 @@ if [ ${RAILS_ENV:-development} != "production" ]; then
   yarn
 fi
 
-echo "starting rails"
-rm -f tmp/pids/server.pid
-bundle exec rails db:create
-bundle exec rails db:migrate
-bundle exec rails solr:init
-bundle exec rails s -b '0.0.0.0'
+if [ ${APP_ROLE:-app} == "sidekiq" ]; then
+    echo "starting sidekiq"
+    /app/bin/sidekiq
+else
+    echo "starting rails"
+    rm -f tmp/pids/server.pid
+    bundle exec rails db:create
+    bundle exec rails db:migrate
+    bundle exec rails solr:init
+    bundle exec rails s -b '0.0.0.0'
+fi
